Add a notification mechanism.
This builds a notification mechanism as a pair of objects:
* A notification that contains the status of the notification; and
* A reusable behaviour.

The reusable behaviour can be scheduled without allocation, and once
a run completes, it either reschedules itself if required, or waits for
a future notification.

Two objects are required as the lifetime management is complex.
The notification object is managed using strong and weak pointers.
While the reusable behaviour must correctly wait for ABA issues on the
MPMC scheduler structures.