Improve perf benchmarking for immutable freeze
This creates a series of benchmarks for immutable data structures.  There are several patterns to increase the coverage of the algorithm from a performance point of view.
* Linked List (Double and Single)
* Balanced binary tree (nullptr for leaves, or root for leaves, i.e. cycle)
* Balanced 4-tree (where each TreeNode is a 4 element linked list)
* Parent pointing tree
* Linked list with lots of backedges to various points.

These cover some fairly deep cyclic structures that are designed to push the corner cases of performance.

The output is generated in a way that is easy to graph.