Work Stealing (#48)
* Add a work stealing benchmark

* Replace implementation of MPMCQ

This replaces the existing queue implementation with a new version that supports dequeue_all.

The dequeue_all will be used in a subsequent PR, which will implement a better work stealing algorithm.

* Add multiple queues per core

This adds a fixed set of queues per core that are used in a round-robin fashion.  This allows more efficient stealing, which will be part of the next commit.

* Fairness

The new stealing mechanism can work with fairness, but works less well.  It is more likely that a steal will take the fairness token, and hence we get some less good behaviour.