Handle Cown duplicates without erasing
The cown duplicates can still be accessed by the behaviour, but the existing code nulls them out to signify they are duplicates.

This PR changes that to use a bit to represent the duplicate cown, and still leave the original pointer intact.  This means that the behaviour body can access the cowns from the behaviour rather than closing over them separately.