Make work stealing scheduler not depend on BoC (#69)
* Remove Detect leaks from scheduler

This is now part of the harness, removing unused functionality.

* Remove barrier from scheduler

The barrier is no longer needed in the scheduler it was part of the complex GC teardown that is not required anymore.

* Separate work scheduler from BoC

The BoC runtime is built ontop of a work stealing scheduler.  The work stealing scheduler does not require the BoC runtime.  The restructures the code to make the work stealing scheduler more standalone.

* Clangformat

* Update CI